(PC-25369) refactor(algolia): fixtures in specific file
