+++ src/features/home/components/modules/video/VideoMonoOfferTile.native.test.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { navigate } from '__mocks__/@react-navigation/native'
 import { SubcategoriesResponseModelv2 } from 'api/gen'
 import { VideoMonoOfferTile } from 'features/home/components/modules/video/VideoMonoOfferTile'
-import { mockedAlgoliaResponse } from 'libs/algolia/__mocks__/mockedAlgoliaResponse'
+import { mockedAlgoliaResponse } from 'libs/algolia/fixtures/algoliaFixtures'
 import { analytics } from 'libs/analytics'
 import { OfferAnalyticsParams } from 'libs/analytics/types'
 import { placeholderData } from 'libs/subcategories/placeholderData'
