+++ src/features/offer/components/OfferTile/OfferTile.native.test.tsx
@@ -5,7 +5,7 @@ import { push } from '__mocks__/@react-navigation/native'
 import { CategoryIdEnum, HomepageLabelNameEnumv2 } from 'api/gen'
 import { Referrals } from 'features/navigation/RootNavigator/types'
 import { PlaylistType } from 'features/offer/enums'
-import { mockedAlgoliaResponse } from 'libs/algolia/__mocks__/mockedAlgoliaResponse'
+import { mockedAlgoliaResponse } from 'libs/algolia/fixtures/algoliaFixtures'
 import { analytics } from 'libs/analytics'
 import { RecommendationApiParams } from 'shared/offer/types'
 import { queryCache, reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
